feat: migrate list item into table component
